refactoring acsp structure
