+++ src/main/java/uk/gov/companieshouse/api/testdata/service/impl/TestDataServiceImpl.java
@@ -1,16 +1,12 @@
 package uk.gov.companieshouse.api.testdata.service.impl;
 
-import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.api.testdata.Application;
 import uk.gov.companieshouse.api.testdata.exception.DataException;
 import uk.gov.companieshouse.api.testdata.exception.NoDataFoundException;
-import uk.gov.companieshouse.api.testdata.model.entity.AcspProfile;
 import uk.gov.companieshouse.api.testdata.model.entity.Postcodes;
-import uk.gov.companieshouse.api.testdata.model.rest.request.AcspMembersRequest;
-import uk.gov.companieshouse.api.testdata.model.rest.request.AcspProfileRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.request.AdminPermissionsRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.request.CertificatesRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.request.CertifiedCopiesRequest;
@@ -20,16 +16,12 @@ import uk.gov.companieshouse.api.testdata.model.rest.request.PenaltyRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.request.TransactionsRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.request.UpdateAccountPenaltiesRequest;
 import uk.gov.companieshouse.api.testdata.model.rest.response.AccountPenaltiesResponse;
-import uk.gov.companieshouse.api.testdata.model.rest.response.AcspMembersResponse;
-import uk.gov.companieshouse.api.testdata.model.rest.response.AcspProfileResponse;
 import uk.gov.companieshouse.api.testdata.model.rest.response.AdminPermissionsResponse;
 import uk.gov.companieshouse.api.testdata.model.rest.response.CertificatesResponse;
 import uk.gov.companieshouse.api.testdata.model.rest.response.CombinedSicActivitiesResponse;
 import uk.gov.companieshouse.api.testdata.model.rest.response.PostcodesResponse;
 import uk.gov.companieshouse.api.testdata.model.rest.response.TransactionsResponse;
-import uk.gov.companieshouse.api.testdata.repository.AcspMembersRepository;
 import uk.gov.companieshouse.api.testdata.service.AccountPenaltiesService;
-import uk.gov.companieshouse.api.testdata.service.AcspProfileService;
 import uk.gov.companieshouse.api.testdata.service.AppealsService;
 import uk.gov.companieshouse.api.testdata.service.DataService;
 import uk.gov.companieshouse.api.testdata.service.ItemGroupsService;
@@ -41,15 +33,12 @@ import uk.gov.companieshouse.logging.LoggerFactory;
 import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class TestDataServiceImpl implements TestDataService {
 
     private static final Logger LOG = LoggerFactory.getLogger(Application.APPLICATION_NAME);
 
-    @Autowired
-    private DataService<AcspMembersResponse, AcspMembersRequest> acspMembersService;
     @Autowired
     private DataService<CertificatesResponse, CertificatesRequest> certificatesService;
     @Autowired
@@ -60,11 +49,7 @@ public class TestDataServiceImpl implements TestDataService {
     @Autowired
     private DataService<CertificatesResponse, MissingImageDeliveriesRequest> missingImageDeliveriesService;
     @Autowired
-    private AcspMembersRepository acspMembersRepository;
-    @Autowired
     private DataService<TransactionsResponse, TransactionsRequest> transactionService;
-    @Autowired
-    private AcspProfileService acspProfileService;
     @Autowired
     AppealsService appealsService;
     @Autowired
@@ -75,78 +60,6 @@ public class TestDataServiceImpl implements TestDataService {
     private DataService<AdminPermissionsResponse, AdminPermissionsRequest> adminPermissionsService;
     @Autowired
     private ItemGroupsService itemGroupsService;
-
-    @Override
-    public AcspMembersResponse createAcspMembersData(final AcspMembersRequest spec) throws DataException {
-        if (spec.getUserId() == null) {
-            throw new DataException("User ID is required to create an ACSP member");
-        }
-
-        var acspProfileSpec = new AcspProfileRequest();
-        if (spec.getAcspProfile() != null) {
-            acspProfileSpec = spec.getAcspProfile();
-        }
-
-        try {
-            var acspProfileData = createAcspProfile(acspProfileSpec);
-            spec.setAcspNumber(acspProfileData.getAcspNumber());
-
-            AcspMembersResponse createdMember = createAcspMember(spec);
-
-            return new AcspMembersResponse(
-                    new ObjectId(createdMember.getAcspMemberId()),
-                    createdMember.getAcspNumber(),
-                    createdMember.getUserId(),
-                    createdMember.getStatus(),
-                    createdMember.getUserRole()
-            );
-        } catch (Exception ex) {
-            throw new DataException(ex);
-        }
-    }
-
-    private AcspProfileResponse createAcspProfile(AcspProfileRequest acspProfileRequest)
-            throws DataException {
-        try {
-            return this.acspProfileService.create(acspProfileRequest);
-        } catch (Exception ex) {
-            throw new DataException("Error creating ACSP profile", ex);
-        }
-    }
-
-    private AcspMembersResponse createAcspMember(AcspMembersRequest spec) throws DataException {
-        try {
-            return this.acspMembersService.create(spec);
-        } catch (Exception ex) {
-            throw new DataException("Error creating ACSP member", ex);
-        }
-    }
-
-    @Override
-    public boolean deleteAcspMembersData(String acspMemberId) throws DataException {
-        List<Exception> suppressedExceptions = new ArrayList<>();
-        try {
-            var maybeMember = acspMembersRepository.findById(acspMemberId);
-            if (maybeMember.isPresent()) {
-                var member = maybeMember.get();
-                String acspNumber = member.getAcspNumber();
-
-                deleteAcspMember(acspMemberId, suppressedExceptions);
-                deleteAcspProfile(acspNumber, suppressedExceptions);
-            } else {
-                return false;
-            }
-        } catch (Exception de) {
-            suppressedExceptions.add(de);
-        }
-
-        if (!suppressedExceptions.isEmpty()) {
-            var ex = new DataException("Error deleting acsp member's data");
-            suppressedExceptions.forEach(ex::addSuppressed);
-            throw ex;
-        }
-        return true;
-    }
 
     @Override
     public CertificatesResponse createCertificatesData(
@@ -352,28 +265,6 @@ public class TestDataServiceImpl implements TestDataService {
             postcodesResponseList.add(postcodeData);
         }
         return postcodesResponseList;
-    }
-
-    public Optional<AcspProfile> getAcspProfileData(String acspNumber)
-            throws NoDataFoundException {
-
-        return acspProfileService.getAcspProfile(acspNumber);
-    }
-
-    private void deleteAcspMember(String acspMemberId, List<Exception> suppressedExceptions) {
-        try {
-            acspMembersService.delete(acspMemberId);
-        } catch (Exception ex) {
-            suppressedExceptions.add(new DataException("Error deleting ACSP member", ex));
-        }
-    }
-
-    private void deleteAcspProfile(String acspNumber, List<Exception> suppressedExceptions) {
-        try {
-            this.acspProfileService.delete(acspNumber);
-        } catch (Exception ex) {
-            suppressedExceptions.add(new DataException("Error deleting ACSP profile", ex));
-        }
     }
 
     public TransactionsResponse createTransactionData(TransactionsRequest transactionsRequest)
